ca bug au niveau des colors

+++ ft_color.c
@@ -6,7 +6,7 @@
 /*   By: mbourdel <mbourdel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/07 15:48:14 by mbourdel          #+#    #+#             */
-/*   Updated: 2015/01/20 12:39:55 by mbourdel         ###   ########.fr       */
+/*   Updated: 2015/04/08 22:59:54 by mbourdel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,31 +14,31 @@
 
 static void			color1(t_pt2d point, t_env *env, unsigned int *color)
 {
-	if (point.memz > (env->value.alt_max * 0.8))
+	if ((float)point.memz > (env->value.alt_max * 0.8))
 		*color = 0xFFFFFF;
-	else if (point.memz > (env->value.alt_max * 0.55)
-				&& (point.memz <= (env->value.alt_max * 0.8)))
+	else if ((float)point.memz > (env->value.alt_max * 0.55)
+				&& ((float)point.memz <= (env->value.alt_max * 0.8)))
 		*color = 0x663300;
 	return ;
 }
 
 static void			color2(t_pt2d point, t_env *env, unsigned int *color)
 {
-	if (point.memz < (env->value.alt_max * 0.55))
+	if ((float)point.memz < (env->value.alt_max * 0.55))
 		*color = 0x996633;
-	if (point.memz < (env->value.alt_max * 0.40))
+	if ((float)point.memz < (env->value.alt_max * 0.40))
 		*color = 0xD1A319;
-	if (point.memz < (env->value.alt_max * 0.30))
+	if ((float)point.memz < (env->value.alt_max * 0.30))
 		*color = 0x669900;
-	if (point.memz < (env->value.alt_max * 0.20))
+	if ((float)point.memz < (env->value.alt_max * 0.20))
 		*color = 0x009900;
-	if (point.memz < (env->value.alt_max * 0.15))
+	if ((float)point.memz < (env->value.alt_max * 0.15))
 		*color = 0x009933;
-	if (point.memz < (env->value.alt_max * 0.10))
+	if ((float)point.memz < (env->value.alt_max * 0.10))
 		*color = 0x006600;
-	if (point.memz < (env->value.alt_max * 0.05))
+	if ((float)point.memz < (env->value.alt_max * 0.05))
 		*color = 0x005200;
-	if (point.memz < (env->value.alt_max * 0.02))
+	if ((float)point.memz < (env->value.alt_max * 0.02))
 		*color = 0x194D19;
 	return ;
 }
@@ -48,7 +48,7 @@ static void			color3(t_pt2d point, t_env *env, unsigned int *color)
 	if (point.memz < 0)
 	{
 		*color = 0x0099FF;
-		if (point.memz < (env->value.alt_min * 0.7))
+		if ((float)point.memz < (env->value.alt_min * 0.7))
 			*color = 0x0000FF;
 	}
 	return ;
